Update to handle UID and DX10His Screen
'ER001 	- Eric Roberts 10/20/2015 - allow passing of UID in option file
"UIDpassed`UID#"
'		- this allows the retrieval the appropriate ICD10 diagnosis record,
based on UID passed
'		- for parmfile DX10VO, this allows a user to select the appropriate
record for display in DX10His
'		- for parmfiles DX10DE and DX10PRM this allows a user to select the
appropriate record for update in DX10His
'		- if UID option is not passed, the default is most recent diagnosis
'		- 10/28/2015, and we are cooing with gas!  Changing the SaveData
function to use the updateClientDx and putClientDx functions coded
'		- in lib_DX10.  Also need to bypass certain edits if UID is passed in